Delete script based installer (#372)
Maintaining 2 installers has become a lot of work. Also, it is becoming difficult to maintain option parity between chart and script installers. With the advent of helm 3, the primary issue with Helm (tiller) is also gone. Helm is much more ubiquitous today than when we started the project. So, we can going to just maintain the helm chart based installer. Those who prefer to not use helm, can continue to use `helm template | kubectl apply -` to install kubed.

Signed-off-by: Tamal Saha <tamal@appscode.com>