Fix wrong annotation used on omni ConfigMap (#482)
The  code to add the annotation to get the omni ConfigMap synced across clusters is wrong. It should be `sync-contexts` instead of just `sync`. As, the line above actually also mentions.

Man did this get me confused into why a Secret of mine was not synced to another cluster. I therefore hope you'll accept this small brush up.